Use real SQS to test that a message is not deleted from the queue if the id_minter cannot insert it into the database
