+++ id_minter/src/test/scala/uk/ac/wellcome/platform/idminter/IdMinterFeatureTest.scala
@@ -3,40 +3,21 @@ package uk.ac.wellcome.platform.idminter
 import com.twitter.finatra.http.EmbeddedHttpServer
 import com.twitter.inject.server.FeatureTestMixin
 import org.scalatest.FunSpec
-import org.scalatest.concurrent.Eventually
 import scalikejdbc.{select, _}
 import uk.ac.wellcome.models._
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.platform.idminter.model.Identifier
-import uk.ac.wellcome.platform.idminter.utils.MysqlLocal
-import uk.ac.wellcome.test.utils.{SNSLocal, SQSLocal}
+import uk.ac.wellcome.platform.idminter.utils.IdMinterTestUtils
 import uk.ac.wellcome.utils.JsonUtil
-
-import scala.collection.JavaConversions._
 
 class IdMinterFeatureTest
     extends FunSpec
-    with FeatureTestMixin
-    with SQSLocal
-    with SNSLocal
-    with MysqlLocal {
+    with FeatureTestMixin with IdMinterTestUtils {
 
-  val ingestorTopicArn: String = createTopicAndReturnArn("test_ingestor")
-  val idMinterQueue: String = createQueueAndReturnUrl("test_id_minter")
-
-  // Setting 1 second timeout for tests, so that test don't have to wait too long to test message deletion
-  sqsClient.setQueueAttributes(idMinterQueue, Map("VisibilityTimeout" -> "1"))
-  override val server: EmbeddedHttpServer = new EmbeddedHttpServer(
-    new Server(),
-    flags = Map(
-      "aws.region" -> "localhost",
-      "aws.sqs.queue.url" -> idMinterQueue,
-      "aws.sqs.waitTime" -> "1",
-      "aws.sns.topic.arn" -> ingestorTopicArn
-    ) ++ snsLocalEndpointFlags ++ sqsLocalFlags ++ mySqlLocalEndpointFlags
-  )
+  override val server: EmbeddedHttpServer = defineServer
 
   private val i = identifiersTable.i
+
   it(
     "should read a work from the SQS queue, generate a canonical ID, save it in dynamoDB and send a message to the SNS topic with the original work and the id") {
     val miroID = "M0001234"
@@ -114,34 +95,9 @@ class IdMinterFeatureTest
     }
   }
 
-  it(
-    "should not delete a message from the sqs queue if it fails processing it") {
+  it("should not delete a message from the sqs queue if it fails processing it") {
     sqsClient.sendMessage(idMinterQueue, "not a json string")
 
-    // After a message is read, it stays invisible for 1 second and then it gets sent again.
-    // So we wait for longer than the visibility timeout and then we assert that it has become
-    // invisible again, which means that the id_minter picked it up again,
-    // and so it wasn't deleted as part of the first run.
-    // TODO Write this test using dead letter queues once https://github.com/adamw/elasticmq/issues/69 is closed
-    Thread.sleep(2000)
-
-    eventually {
-      sqsClient
-        .getQueueAttributes(idMinterQueue,
-                            List("ApproximateNumberOfMessagesNotVisible"))
-        .getAttributes
-        .get("ApproximateNumberOfMessagesNotVisible") shouldBe "1"
-    }
-  }
-
-  private def generateSqsMessage(MiroID: String) = {
-    val work = Work(identifiers =
-                      List(SourceIdentifier("Miro", "MiroID", MiroID)),
-                    label = "some label")
-    SQSMessage(Some("subject"),
-               JsonUtil.toJson(work).get,
-               "topic",
-               "messageType",
-               "timestamp")
+    assertMessageIsNotDeleted()
   }
 }
